requirements.txt
pillow
nuitka